requirements.txt
tuyapy
flask
paho-mqtt
requests
